+++ lib/src/keys.dart
@@ -1,8 +1,11 @@
+import 'dart:typed_data';
+
+import 'package:basic_utils/basic_utils.dart';
+import 'package:dart_jsonwebtoken/src/utils.dart';
+import 'package:ed25519_edwards/ed25519_edwards.dart' as ed;
 import 'package:pointycastle/pointycastle.dart' as pc;
 
-import 'package:ed25519_edwards/ed25519_edwards.dart' as ed;
 import 'exceptions.dart';
-import 'crypto_utils.dart';
 
 abstract class JWTKey {}
 
@@ -27,6 +30,8 @@ class RSAPrivateKey extends JWTKey {
 
   RSAPrivateKey.raw(pc.RSAPrivateKey _key) : key = _key;
   RSAPrivateKey.clone(RSAPrivateKey _key) : key = _key.key;
+  RSAPrivateKey.bytes(Uint8List bytes)
+      : key = CryptoUtils.rsaPrivateKeyFromDERBytes(bytes);
 }
 
 /// For RSA algorithm, in verify method
@@ -43,6 +48,26 @@ class RSAPublicKey extends JWTKey {
 
   RSAPublicKey.raw(pc.RSAPublicKey _key) : key = _key;
   RSAPublicKey.clone(RSAPublicKey _key) : key = _key.key;
+  RSAPublicKey.bytes(Uint8List bytes) {
+    try {
+      key = CryptoUtils.rsaPublicKeyFromDERBytesPkcs1(bytes);
+    } catch (_) {
+      key = CryptoUtils.rsaPublicKeyFromDERBytes(bytes);
+    }
+  }
+  RSAPublicKey.cert(String pem) {
+    final x509 = X509Utils.x509CertificateFromPem(pem);
+    final bytes = x509.tbsCertificate?.subjectPublicKeyInfo.bytes;
+    if (bytes == null) {
+      throw JWTParseException('x509 Certificate parsing failed');
+    }
+
+    try {
+      key = CryptoUtils.rsaPublicKeyFromDERBytesPkcs1(hexToUint8List(bytes));
+    } catch (_) {
+      key = CryptoUtils.rsaPublicKeyFromDERBytes(hexToUint8List(bytes));
+    }
+  }
 }
 
 /// For ECDSA algorithm, in sign method
@@ -75,6 +100,8 @@ class ECPrivateKey extends JWTKey {
   ECPrivateKey.clone(ECPrivateKey _key)
       : key = _key.key,
         size = _key.size;
+  ECPrivateKey.bytes(Uint8List bytes)
+      : key = CryptoUtils.ecPrivateKeyFromDerBytes(bytes);
 }
 
 /// For ECDSA algorithm, in verify method
@@ -87,6 +114,8 @@ class ECPublicKey extends JWTKey {
 
   ECPublicKey.raw(pc.ECPublicKey _key) : key = _key;
   ECPublicKey.clone(ECPublicKey _key) : key = _key.key;
+  ECPublicKey.bytes(Uint8List bytes)
+      : key = CryptoUtils.ecPublicKeyFromDerBytes(bytes);
 }
 
 /// For EdDSA algorithm, in sign method