+++ lib/src/utils.dart
@@ -89,3 +89,19 @@ BigInt bigIntFromBytes(Uint8List bytes) {
 
   return bytes.fold(BigInt.zero, (a, b) => a * _b256 + BigInt.from(b));
 }
+
+Uint8List hexToUint8List(String hex) {
+  if (hex.length % 2 != 0) {
+    throw 'Odd number of hex digits';
+  }
+  var l = hex.length ~/ 2;
+  var result = Uint8List(l);
+  for (var i = 0; i < l; ++i) {
+    var x = int.parse(hex.substring(2 * i, 2 * (i + 1)), radix: 16);
+    if (x.isNaN) {
+      throw 'Expected hex string';
+    }
+    result[i] = x;
+  }
+  return result;
+}